Docker/app/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
psycopg2-binary==2.9.9